+++ website/src/options/defaults.ts
@@ -1,31 +1,33 @@
 export default {
+  background: {
+    type: 'color',
+    defaultValue: '#ffffff',
+    alias: 'b',
+  },
   radius: {
-    type: "range",
+    type: 'range',
     values: [0, 50],
     defaultValue: 0,
-    alias: "r",
+    alias: 'r',
   },
-  margin: {
-    type: "range",
-    values: [0, 25],
-    defaultValue: 0,
-    alias: "m",
-  },
-  background: {
-    type: "color",
-    defaultValue: "#ffffff",
-    alias: "b",
-  },
-  width: {
-    type: "number",
+  size: {
+    type: 'number',
     values: [0, 2048],
     defaultValue: 0,
-    alias: "w",
   },
-  height: {
-    type: "number",
-    values: [0, 2048],
+  scale: {
+    type: 'range',
+    values: [50, 150],
+    defaultValue: 100,
+  },
+  flip: {
+    type: 'switch',
+    values: [0, 1],
     defaultValue: 0,
-    alias: "h",
+  },
+  rotate: {
+    type: 'range',
+    values: [0, 360],
+    defaultValue: 0,
   },
 };