+++ CHANGELOG.md
@@ -6,6 +6,17 @@
 
 - `@dicebear/croodles` New Avatar style by vijay verma.
 - `@dicebear/croodles-neutral` New Avatar style by vijay verma.
+- `@dicebear/avatars` New `rotate` option.
+- `@dicebear/avatars` New `flip` option.
+- `@dicebear/avatars` New `size` option as replacement for `width` and `height` options.
+
+### Deprecated
+
+- `@dicebear/avatars` `width` and `height` options are deprecated. Use `size` instead.
+- `@dicebear/avatars` `margin` option is deprecated. Use `scale` instead.
+- `@dicebear/avatars` SVG Util `addMargin` is deprecated. Use `addScale` instead.
+- `@dicebear/avatars` SVG Util `addRadius` is deprecated. Use `addViewboxMask` instead.
+- `@dicebear/avatars` SVG Util `addBackgroundColor` is deprecated.
 
 ## [4.7.4] - 2021-08-25